backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.9.0
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
deprecated==1.2.18
distro==1.9.0
elevenlabs==2.5.0
eval-type-backport==0.2.2
fastapi==0.115.6
filelock==3.18.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.10
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-speech==2.28.0
google-cloud-texttospeech==2.18.0
google-generativeai==0.8.3
googleapis-common-protos==1.59.1
griffe==1.7.3
groq==0.30.0
grpcio==1.73.1
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-metadata==8.7.0
jinja2==3.1.6
jiter==0.10.0
json-repair==0.47.7
llvmlite==0.44.0
logfire-api==3.24.2
markupsafe==3.0.2
mistralai==1.9.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.0
openai==1.57.2
openai-whisper==20250625
opentelemetry-api==1.35.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.56b0
opentelemetry-instrumentation-asgi==0.56b0
opentelemetry-instrumentation-fastapi==0.56b0
opentelemetry-instrumentation-httpx==0.56b0
opentelemetry-instrumentation-logging==0.56b0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opentelemetry-util-http==0.56b0
packaging==25.0
pip==25.1.1
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.10.3
pydantic-ai==0.0.14
pydantic-ai-slim==0.0.14
pydantic-core==2.27.1
pydantic-graph==0.1.7
pydantic-settings==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.14.0
thrift==0.22.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.1.0
websockets==13.1
wrapt==1.17.2
zipp==3.23.0
